sample/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
biopython==1.78
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-datetime-widget2==0.9.5
django-excel==0.0.10
django-filter==2.4.0
django-import-export==2.4.0
django-tables2==2.3.4
et-xmlfile==1.0.1
filelock==3.0.12
jdcal==1.4.1
lml==0.1.0
MarkupPy==1.14
networkx==2.5
numpy==1.19.3
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
prettytable==2.0.0
pydna==3.1.3
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
texttable==1.6.3
urllib3==1.26.2
virtualenv==20.2.2
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
uWSGI==2.0.19.1
